refactor: sp 수정api 검증 로직 추가
- thumbnailImg 존재하는 경우에만 ThumbnailImgKey 수정 로직 수행

+++ src/main/java/com/pitchain/service/SpService.java
@@ -120,11 +120,13 @@ public class SpService {
 
         validateSpOwner(sp, company);
 
-        s3Service.deleteImg(sp.getThumbnailImgKey());
-        String thumbnailImgKey = s3Service.uploadFile(thumbnailImg, S3UploadTarget.COMPANY_THUMBNAIL);
+        if (thumbnailImg != null) {
+            s3Service.deleteImg(sp.getThumbnailImgKey());
+            String thumbnailImgKey = s3Service.uploadFile(thumbnailImg, S3UploadTarget.COMPANY_THUMBNAIL);
+            sp.updateThumbnailImgKey(thumbnailImgKey);
+        }
 
-        Sp updateSp = Sp.of(sp.getBm(), thumbnailImgKey, name);
-        sp.update(updateSp);
+        sp.update(name);
     }
 
     public void deleteSp(MemberDetails memberDetails, Long spId) {
